Fixed waiting cursor
The login screen is shown while we still wait for the PDR to respond.
However, the proxy has set the cursor to a waiting state.
Consequently, we would see a waiting cursor while logging in
and this is confusing.
We now restore the cursor shape on showing the login screen.

We've also made sure that the entire viewport is used, so the waiting
cursor is visible everywhere.